+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support rare PHP installations without `sockets` extension
 - Fixed rare fatal error in `show_error_and_die()`
 - Fixed deprecation notice warning in `sanitize_key_part()`
+- Added timeouts to diagnostics
 
 ## 2.5.0